+++ tests/unit/test_parameters/test_parameter_values.py
@@ -2,10 +2,10 @@
 # Tests for the Base Parameter Values class
 #
 import pybamm
-
-import unittest
+import os
 import numpy as np
 
+import unittest
 import tests.shared as shared
 
 
@@ -25,7 +25,7 @@ class TestParameterValues(unittest.TestCase):
             values="input/parameters/lithium-ion/cathodes/lico2_Marquis2019/"
             + "parameters.csv"
         )
-        self.assertEqual(param["Reference temperature [K]"], "298.15")
+        self.assertEqual(param["Reference temperature [K]"], 298.15)
 
         # values vs chemistry
         with self.assertRaisesRegex(
@@ -293,6 +293,61 @@ class TestParameterValues(unittest.TestCase):
         processed_func = parameter_values.process_symbol(func)
         self.assertIsInstance(processed_func, pybamm.Function)
         self.assertEqual(processed_func.evaluate(), 3)
+
+    def test_process_interpolant(self):
+        x = np.linspace(0, 10)[:, np.newaxis]
+        data = np.hstack([x, 2 * x])
+        parameter_values = pybamm.ParameterValues(
+            {"a": 3.01, "Diffusivity": ("times two", data)}
+        )
+
+        a = pybamm.Parameter("a")
+        func = pybamm.FunctionParameter("Diffusivity", a)
+
+        processed_func = parameter_values.process_symbol(func)
+        self.assertIsInstance(processed_func, pybamm.Interpolant)
+        self.assertEqual(processed_func.evaluate(), 6.02)
+
+        # process differentiated function parameter
+        diff_func = func.diff(a)
+        processed_diff_func = parameter_values.process_symbol(diff_func)
+        self.assertEqual(processed_diff_func.evaluate(), 2)
+
+    def test_interpolant_against_function(self):
+        parameter_values = pybamm.ParameterValues({"a": 0.6})
+        parameter_values.update(
+            {
+                "function": "[function]lico2_ocp_Dualfoil1998",
+                "interpolation": "[data]lico2_data_example",
+            },
+            path=os.path.join(
+                pybamm.root_dir(),
+                "input",
+                "parameters",
+                "lithium-ion",
+                "cathodes",
+                "lico2_Marquis2019",
+            ),
+        )
+
+        a = pybamm.Parameter("a")
+        func = pybamm.FunctionParameter("function", a)
+        interp = pybamm.FunctionParameter("interpolation", a)
+
+        processed_func = parameter_values.process_symbol(func)
+        processed_interp = parameter_values.process_symbol(interp)
+        np.testing.assert_array_almost_equal(
+            processed_func.evaluate(), processed_interp.evaluate(), decimal=4
+        )
+
+        # process differentiated function parameter
+        diff_func = func.diff(a)
+        diff_interp = interp.diff(a)
+        processed_diff_func = parameter_values.process_symbol(diff_func)
+        processed_diff_interp = parameter_values.process_symbol(diff_interp)
+        np.testing.assert_array_almost_equal(
+            processed_diff_func.evaluate(), processed_diff_interp.evaluate(), decimal=2
+        )
 
     def test_process_complex_expression(self):
         var1 = pybamm.Variable("var1")