+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Features
 
+- Add `Interpolant` class to interpolate experimental data (e.g. OCP curves) (#661)
 - Allow parameters to be set by material or by specifying a particular paper (#647)
 - Set relative and absolute tolerances independently in solvers (#645)